feat(testing): promote ScriptedProvider to public lovia.testing
The scripted provider moves from tests/ to a public home so user eval
suites and example code can run offline. tests/scripted_provider.py
stays as a re-export shim; existing test imports are unchanged.
